refactor: fix ESLint max-lines and complexity warnings
- Extract TransformMatchingShared.ts from TransformMatching.ts
- Extract VMobjectCurveUtils.ts from VMobject.ts (curve utilities)
- Extract VMobjectStrokeGeometry.ts from VMobjectGeometry.ts
- Reduce alignCompoundPathsForTransform complexity by extracting validation helper
- Update tests to import curve utilities from VMobjectCurveUtils
- Remove circular dependency by not re-exporting from VMobject.ts
